By Devinda - 2nd update
Updated the Navbar into a component
